+++ sbuildr/generator/__init__.py
@@ -1,0 +1,1 @@
+from sbuildr.generator.rbuild import RBuildGenerator