+++ sbuildr/graph/node.py
@@ -1,6 +1,6 @@
-from srbuild.tools import compiler, linker
-from srbuild.tools.flags import BuildFlags
-from srbuild.logger import G_LOGGER
+from sbuildr.tools import compiler, linker
+from sbuildr.tools.flags import BuildFlags
+from sbuildr.logger import G_LOGGER
 
 from typing import List
 import os