+++ srbuild/generator/__init__.py
@@ -1,1 +1,0 @@
-from srbuild.generator.rbuild import RBuildGenerator