+++ tests/minimal_project/build.py
@@ -3,13 +3,13 @@
 import os
 import sys
 CURDIR = os.path.dirname(__file__)
-SRBUILD_DIR = os.path.abspath(os.path.join(CURDIR, os.pardir, os.pardir))
-sys.path.insert(0, SRBUILD_DIR)
+SBUILDR_DIR = os.path.abspath(os.path.join(CURDIR, os.pardir, os.pardir))
+sys.path.insert(0, SBUILDR_DIR)
 
 # Normal build files would start here:
-import srbuild
+import sbuildr
 
-project = srbuild.Project()
+project = sbuildr.Project()
 libmath = project.library("math", sources=["factorial.cpp", "fibonacci.cpp"], libs=["stdc++"])
 libtest = project.library("test", sources=["factorial.cpp", "fibonacci.cpp"], libs=["stdc++"])
 install_dir = os.path.join(CURDIR, "build")
@@ -18,4 +18,4 @@ project.install("utils.hpp", dir=install_dir)
 test = project.test("test", sources=["test.cpp"], libs=["stdc++", libtest])
 test = project.executable("test_e", sources=["test.cpp"], libs=["stdc++", libtest])
 
-srbuild.cli(project)
+sbuildr.cli(project)